browserstack-local run-with npm script
Issue with using start/stop in npm pre/post scripts is if the target npm command errors out,
the binary isn't stopped since it runs in daemon mode. If we (cmd || browserstack-local stop),
the proper exit code isn't propagated for the CI run to be marked as failed.

run-with wraps the npm test command and copies over the child process exit code.
TODO: test cross-platform